Kyle use cli branch, grid uses cli now

+++ generator/src/main/java/ca/mcmaster/cas/se2aa4/a2/generator/MySegment.java
@@ -114,10 +114,10 @@ public class MySegment {
     }
 
     /**
-     * Sets the thickness of this vertex.
+     * Sets the thickness of this segment.
      * @param thickness int value of thickness wanted
      */
-    public void setThickness(int thickness){
+    public void setThickness(float thickness){
         Property thick = Property.newBuilder().setKey("thickness").setValue("" + thickness).build();
 
         String val = PropertyManager.getProperty(this.getPropertiesList(), "thickness");
